Separate base Runner class and GCP subclasses into separate files
